backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv==1.2.1
motor==3.3.1
pydantic==2.12.5
email-validator==2.3.0
PyJWT==2.11.0
passlib[bcrypt]==1.7.4
razorpay==2.0.0
